warning removed
Rev: src/las.c:1.2
Rev: src/lex.c:1.2

+++ src/las.c
@@ -166,7 +166,7 @@ node *mknode(short token,node *a,node *b)
   case F_APPLY:
     if(a && a->token == F_CONSTANT &&
        a->u.sval.type == T_FUNCTION &&
-       a->u.sval.subtype == -1)
+       (short)a->u.sval.subtype == -1)
     {
       res->node_info |= a->u.sval.u.efun->flags;
     }else{
@@ -414,7 +414,7 @@ node *mksvaluenode(struct svalue *s)
 
   case T_FUNCTION:
   {
-    if(s->subtype != -1)
+    if((short)s->subtype != -1)
     {
       if(s->u.object == &fake_object)
 	return mkidentifiernode(s->subtype);
@@ -1167,7 +1167,7 @@ static void optimize(node *n)
     case F_APPLY:
       if(CAR(n)->token == F_CONSTANT &&
 	 CAR(n)->u.sval.type == T_FUNCTION &&
-	 CAR(n)->u.sval.subtype == -1 && /* driver fun? */
+	 (short)CAR(n)->u.sval.subtype == -1 && /* driver fun? */
 	 CAR(n)->u.sval.u.efun->optimize)
       {
 	if(tmp1=CAR(n)->u.sval.u.efun->optimize(n))
@@ -1372,7 +1372,7 @@ static void optimize(node *n)
 	   CAR(opnode) &&
 	   CAR(opnode)->token == F_CONSTANT &&
 	   CAR(opnode)->u.sval.type == T_FUNCTION &&
-	   CAR(opnode)->u.sval.subtype == -1)
+	   (short)CAR(opnode)->u.sval.subtype == -1)
 	{
 	  if(CAR(opnode)->u.sval.u.efun->function == f_gt)
 	    oper=F_GT;
